pyproject.toml
[tool.poetry]
name = "jenkins_repl"
version = "0.1.0"
description = ""
authors = ["James Hounshell <jameshounshell@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-jenkins = "^1.7.0"
Jinja2 = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-watch = "^4.2.0"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
